Error when installing new kunena template
Joomla\Filesystem\Folder::delete: Path is not a folder. Path: [ROOT]/components/com_kunena/template/mytemplate

+++ src/admin/src/Controller/TemplatesController.php
@@ -146,7 +146,9 @@ class TemplatesController extends KunenaController
                                     File::move($dest . $template->directory . '/assets/css/custom.css', $tmpKunena . $template->sourcedir . '/assets/css/custom.css');
                                 }
 
-                                Folder::delete($dest . $template->directory);
+                                if (is_dir($dest . $template->directory)) {
+                                    Folder::delete($dest . $template->directory);
+                                }
 
                                 $success = Folder::move($tmpKunena . $template->sourcedir, $dest . $template->directory);
